requirements.txt
# OF Assistant Bot - Оптимизированные зависимости
# Совместимо с Python 3.11+

# Основные зависимости
pyTelegramBotAPI==4.21.0
# DeepSeek API использует OpenAI-совместимый интерфейс - можно использовать openai SDK
# или просто aiohttp для прямых API вызовов
aiofiles==23.2.1
psutil==5.9.8
python-dotenv==1.0.0
loguru==0.7.2

# Асинхронность
aiohttp==3.9.3

# Утилиты
python-dateutil==2.8.2
cachetools==5.3.2

# Валидация
pydantic==2.6.1

# Опциональные для разработки (раскомментируйте при необходимости)
# pytest==8.1.1
# pytest-asyncio==0.21.1
# pytest-cov==4.1.0 